backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
soupsieve==2.8.1
starlette==0.49.3
sympy==1.14.0
torch==2.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
python-dotenv==1.0.1
uvicorn==0.39.0
gunicorn==23.0.0
google-api-python-client>=2.110.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# AI/ML - Mistral Integration
mistralai>=1.0.0
tiktoken>=0.5.0
aiohttp>=3.9.0

# Real-time Communication
python-socketio>=5.10.0

# JWT & Security
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0

# Redis for WebSocket scaling
redis>=5.0.0

# Structured logging
python-json-logger>=2.0.0

# Token encryption
cryptography>=41.0.0
